simplify the logic of the test suite

+++ ci/script.sh
@@ -13,9 +13,11 @@ main() {
 
     cargo install --path .
 
-    # `cross build` test for targets where `std` is not available
+    export QEMU_STRACE=1
+
+    # `cross run` test for thumb targets
     case $TARGET in
-        thumbv*-none-eabi*)
+        thumb*-none-eabi*)
             td=$(mktemp -d)
 
             git clone \
@@ -28,49 +30,33 @@ main() {
             popd
 
             rm -rf $td
-
-            return
         ;;
     esac
 
     # `cross build` test for targets where `std` is not available
-    case $TARGET in
-        sparc64-* | \
-            thumbv*-none-eabi* | \
-            x86_64-unknown-dragonfly)
-            td=$(mktemp -d)
+    if [ -z $STD ]; then
+        td=$(mktemp -d)
 
-            git clone \
-                --depth 1 \
-                --recursive \
-                https://github.com/rust-lang-nursery/compiler-builtins $td
+        git clone \
+            --depth 1 \
+            --recursive \
+            https://github.com/rust-lang-nursery/compiler-builtins $td
 
-            pushd $td
-            cat > Cross.toml <<EOF
+        pushd $td
+        cat > Cross.toml <<EOF
 [build]
 xargo = true
 EOF
-            cross build --features c --lib --target $TARGET
-            popd
-
-            rm -rf $td
-
-            return
-            ;;
-    esac
-
-    # `cross build` test for the other targets
-    if [ $TARGET = i686-apple-darwin ] || [ $TARGET = i686-unknown-linux-musl ]; then
-        td=$(mktemp -d)
-
-        git clone --depth 1 https://github.com/japaric/xargo $td
-
-        pushd $td
-        cross build --target $TARGET
+        cross build --features c --lib --target $TARGET
         popd
 
         rm -rf $td
-    else
+
+        return
+    fi
+
+    # `cross build` test for the other targets
+    if [ $OPENSSL ]; then
         td=$(mktemp -d)
 
         git clone --depth 1 https://github.com/rust-lang/cargo $td
@@ -80,24 +66,21 @@ EOF
         popd
 
         rm -rf $td
+    else
+        td=$(mktemp -d)
+
+        git clone --depth 1 https://github.com/japaric/xargo $td
+
+        pushd $td
+        cross build --target $TARGET
+        popd
+
+        rm -rf $td
     fi
 
-    # `cross test` / `cross run` test for the other targets
-    # NOTE(s390x) japaric/cross#3
-    # NOTE(*-musl) can't test compiler-builtins because that crate needs
-    # cdylibs and musl targets don't support cdylibs
-    # NOTE(*-*bsd) no `cross test` support for BSD targets
-    # NOTE(sparc64-*) no `std` available
-    case $TARGET in
-        i686-unknown-freebsd | \
-            i686-unknown-linux-musl | \
-            s390x-unknown-linux-gnu | \
-            sparc64-unknown-linux-gnu | \
-            x86_64-unknown-freebsd | \
-            x86_64-unknown-linux-musl | \
-            x86_64-unknown-netbsd)
-        ;;
-        *)
+    if [ $RUN ]; then
+        # `cross test` test
+        if [ $DYLIB ]; then
             td=$(mktemp -d)
 
             git clone \
@@ -113,43 +96,39 @@ EOF
             popd
 
             rm -rf $td
+        fi
 
-            td=$(mktemp -d)
+        # `cross run` test
+        td=$(mktemp -d)
 
-            cargo init --bin --name hello $td
+        cargo init --bin --name hello $td
 
-            pushd $td
-            QEMU_STRACE=1 cross run --target $TARGET
-            popd
+        pushd $td
+        cross run --target $TARGET
+        popd
 
-            rm -rf $td
-        ;;
-    esac
+        rm -rf $td
+    fi
 
     # Test C++ support
-    case $TARGET in
-        *-unknown-*bsd | \
-            *-unknown-linux-musl)
-            ;;
-        *)
-            td=$(mktemp -d)
+    if [ $CPP ]; then
+        td=$(mktemp -d)
 
-            git clone --depth 1 https://github.com/japaric/hellopp $td
+        git clone --depth 1 https://github.com/japaric/hellopp $td
 
-            pushd $td
-            if [ $TARGET = s390x-unknown-linux-gnu ]; then
-                cross build --target $TARGET
-            else
-                cross run --target $TARGET
-            fi
-            popd
+        pushd $td
+        if [ $TARGET = s390x-unknown-linux-gnu ]; then
+            cross build --target $TARGET
+        else
+            cross run --target $TARGET
+        fi
+        popd
 
-            rm -rf $td
-            ;;
-    esac
+        rm -rf $td
+    fi
 
     # Test openssl compatibility
-    if [ $TRAVIS_OS_NAME = linux ] && [ ! -z "$OPENSSL_INCLUDE_PATH"] && [ ! -z "$OPENSSL_LIB_PATH" ]; then
+    if [ $OPENSSL ]; then
         td=$(mktemp -d)
 
         pushd $td
